Fix squashed commit messages
In https://github.com/MercuryTechnologies/bors-ng/pull/1 I added
support for squash merging PRs in such a way that they are marked
merged, but in the course of doing so I broke the generated commit
message.  Specifically, the things I broke were:

- The commit message no longer includes the original PR number

- The commit message doesn't respect the `cut_body_after` option

- The commit message uses the default message generated by GitHub

  … which will typically be the ugly list of commit titles/descriptions
  rather than the nicer message based off of the PR description.

All three of these things broke for the same reason: `bors-ng`
formats the commit title/message when generate commits for the
staging branch, but we no longer reuse that staging branch when the
batch is accepted.  Instead, the change from
https://github.com/MercuryTechnologies/bors-ng/pull/1 just squash
merges the original PR without massaging the title or description,
which leads to these issues.

The fix is to reuse the same logic that `bors-ng` created for
generating commit messages on the staging branch to also generate
commit messages when squash merging the original PR.  This should
fix all of the above issues.